requirements.txt
Django>=4.2
djangorestframework
gunicorn
whitenoise
python-dotenv
dj-database-url
psycopg2-binary
numpy
pandas
scikit-learn
